closed realms and move open realms to only providers
